fix(features): rework features so that we can test awc_client
* Cargo features are additive in nature and default features are
included when specifying additional features. There is no ability to
remove a default. Our old logic exposed this bug when we tried to
compile with  the `awc_client` feature. We had feature `hyper_client`
as the default to make it easy to build without specifying which http
server implementation to use; which inadvertently set both features to
enabled, when we choose the alternative, `awc_client`. When choosing
feature = `awc_client`; `hyper_client` is also present, as it is the
default. Therefore, we evaluate the not `awc_client` feature to enable
the default behaviour.
* add benchmark reports showing awc_client is superior to hyper_client
* benchmark script now takes a features parameter
* benchmark script opens browser when completed