fix: revert 1.6.0 3rd attempt (#83)
* revert: release 1.6.0

This reverts commit 7046f63ed1873414e37e4a7a7c0046ce62438999.

* fix: remove skip to allow tags workflow to trigger

When the tagged commit is added to the repo, we need the docker tags workflow to
create the container images. If we set the ci to skip, then the workflow is not
triggered.

* ci: use a pat to create the release commit and tags

The GITHUB_TOKEN does not allow triggering of workflows. The persisted credentials
from the actions/checkout overrides the env tokens used by semantic release and
therefore these credentials must be purged.

This allows the PAT to create the new version commit and tag to allow the docker
workflow to create the docker images witht he correct release tag versions.

* chore: rename the release drafter to show upcoming commits in a new release

* ci: formatting and do not build images in forked repos

* chore: add copyright and formatting

* ci: continue pr cleanup if jobs fail

Allow the workflow to be cancelled by evaluating failure